+++ NEWS.md
@@ -1,4 +1,6 @@
 # tidypopgen dev
+* implement our own upset plot function to avoid dependency on `UpSetR`
+  package, which is no longer maintained
 * fix reading of VCFs when genotype separator symbol is present in the
   sample information
 * fix `loci_` functions output when calculating across groups for a single loci